Hotel_Booking_Analytics/Hotel_App/requirements.txt
pandas
numpy
torch
transformers
sentence-transformers
langchain_chroma 
langchain_ollama 
langchain_core
langchain_groq
langchain_community
flask
uvicorn
requests
load_dotenv
langchain_ollama

